Move OsgiDelegatedClassLoade into :libs:utilities
