+++ libs/messaging/kafka-message-bus-impl/build.gradle
@@ -11,10 +11,10 @@ dependencies {
     
     implementation project(":libs:messaging:messaging")
     implementation project(":libs:messaging:message-bus")
-    implementation project(":libs:messaging:kafka-utils")
     implementation project(":libs:schema-registry:schema-registry")
     implementation project(":libs:configuration:configuration-core")
     implementation project(":libs:configuration:configuration-merger")
+    implementation project(":libs:utilities")
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     implementation "net.corda:corda-avro-schema"
