--- libs/messaging/kafka-utils/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'corda.common-library'
-    id 'corda.common-publishing'
-}
-
-description 'Kafka Utilities'
-
-dependencies {
-    compileOnly 'org.osgi:org.osgi.service.component.annotations'
-    compileOnly 'org.osgi:osgi.core'
-
-    implementation platform("net.corda:corda-api:$cordaApiVersion")
-    implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
-
-    testImplementation "org.mockito:mockito-core:$mockitoVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-
-    testRuntimeOnly 'org.osgi:osgi.core'
-}
