--- libs/messaging/kafka-utils/src/main/java/net.corda.messaging.kafka.utils/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.messaging.kafka.utils;
-
-import org.osgi.annotation.bundle.Export;
