+++ settings.gradle
@@ -267,7 +267,6 @@ include 'libs:messaging:db-message-bus-impl'
 include 'libs:messaging:db-topic-admin-impl'
 include 'libs:messaging:kafka-message-bus-impl'
 include 'libs:messaging:kafka-topic-admin-impl'
-include 'libs:messaging:kafka-utils'
 include 'libs:messaging:message-bus'
 include 'libs:messaging:messaging'
 include 'libs:messaging:messaging-impl'
